feat: 날씨 정보 코드 개선
- 초단기실황 API 추가 : 정확한 실시간 기온 및 날씨 상태 출력을 위해 초단기실황 API를 호출하도록 코드 수정
- Zustand 상태 관리 : 날씨 정보와 위치 데이터를 전역 상태로 관리하기 위해 코드 추가
- 로컬 스토리지 저장 : 날씨 데이터를 로컬 스토리지에 저장
  1. 현재 위치 (주소): Kakao API에서 받아온 시, 구, 동 데이터를 가공
  2. 현재 위치와 현재 시간 기준의 TMP(기온) 데이터 저장
  3. 현재 위치와 현재 시간 기준의 날씨
    * 날씨 아이콘은 시간에 맞춰 맑음(낮) 등으로 구분하여 저장(skyCondition)
    * 텍스트는 날씨 상태를 낮과 밤 구분 없이 저장(weatherText)
  4. 오늘부터 3일 뒤까지의 날씨:
    * 3시간 간격의 예보 데이터를 모아서 일자별로 평균 값을 계산해 저장
  5. 타임스탬프 함께 저장 : 기상청 API가 업데이트 될 때를 대비
    * API 데이터를 우선적으로 사용하고, 성공 시 로컬 스토리지에 저장하여 캐싱
    * API 호출 실패 시 로컬 스토리지의 데이터를 임시로 사용
    * 데이터를 시간 제한을 두고 로컬 스토리지에 저장
    * 일정 시간이 지나면 실시간 데이터를 다시 받아오도록 처리
